# Optimize ActiveSupport::NumericWithFormat#to_s