--- a/activesupport/lib/active_support/core_ext/numeric/conversions.rb
+++ b/activesupport/lib/active_support/core_ext/numeric/conversions.rb
@@ -107,9 +107,9 @@ module ActiveSupport
     #                   separator: ',',
     #                   significant: false)                   # => "1,2 Million"
     def to_s(format = nil, options = nil)
+      return super() if format.nil?
+
       case format
-      when nil
-        super()
       when Integer, String
         super(format)
       when :phone